requirements.txt
numpy
torch
timm
pandas
albumentations
scikit-learn
tqdm
pillow
scipy

